refactor: make the comment stack subject-agnostic
Comments hang off proposals only, but sessions and profiles are next.
Split the repository into a core that works on any comment plus one
repository per subject, and lift the proposal comment UI into a shared
CommentsSection that takes what differs — the create action and the
permalink — as props.

issue #461
