+++ db/container.ts
@@ -4,7 +4,10 @@ import path from "path";
 import * as schema from "./schema";
 import { resolveDbPath, runMigrations } from "./migrate";
 import { SqliteAuthCodesRepository } from "./repositories/sqlite/auth-codes";
-import { SqliteCommentsRepository } from "./repositories/sqlite/comments";
+import {
+  SqliteCommentsRepository,
+  sqliteSubjectCommentsRepository,
+} from "./repositories/sqlite/comments";
 import { SqliteDaysRepository } from "./repositories/sqlite/days";
 import { SqliteEventsRepository } from "./repositories/sqlite/events";
 import { SqliteGuestsRepository } from "./repositories/sqlite/guests";
@@ -25,12 +28,15 @@ import type {
   SettingsRepository,
   SessionProposalsRepository,
   SessionsRepository,
+  SubjectCommentsRepository,
   VotesRepository,
 } from "./repositories/interfaces";
 
 export type Repositories = {
   authCodes: AuthCodesRepository;
+  /** Scope-agnostic comment operations (find, edit, like, delete). */
   comments: CommentsRepository;
+  proposalComments: SubjectCommentsRepository;
   days: DaysRepository;
   events: EventsRepository;
   guests: GuestsRepository;
@@ -50,6 +56,7 @@ function buildRepositories(sqlite: Database.Database): Repositories {
   return {
     authCodes: new SqliteAuthCodesRepository(db),
     comments: new SqliteCommentsRepository(db),
+    proposalComments: sqliteSubjectCommentsRepository(db, "proposal"),
     days: new SqliteDaysRepository(db),
     events: new SqliteEventsRepository(db),
     guests: new SqliteGuestsRepository(db),
