+++ db/repositories/sqlite/comments.ts
@@ -1,9 +1,14 @@
 import { and, asc, eq, inArray, sql, type SQL } from "drizzle-orm";
 import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
-import type { SQLiteTable } from "drizzle-orm/sqlite-core";
+import type { AnySQLiteColumn, SQLiteTable } from "drizzle-orm/sqlite-core";
 import { nanoid } from "nanoid";
 import * as schema from "../../schema";
-import type { Comment, CommentLiker, CommentsRepository } from "../interfaces";
+import type {
+  Comment,
+  CommentLiker,
+  CommentsRepository,
+  SubjectCommentsRepository,
+} from "../interfaces";
 
 type DB = BetterSQLite3Database<typeof schema>;
 
@@ -18,136 +23,120 @@ type CommentRow = {
   authorName: string | null;
 };
 
+const commentColumns = {
+  id: schema.comments.id,
+  parentId: schema.comments.parentId,
+  body: schema.comments.body,
+  deleted: schema.comments.deleted,
+  createdTime: schema.comments.createdTime,
+  editedTime: schema.comments.editedTime,
+  authorId: schema.guests.id,
+  authorName: schema.guests.name,
+};
+
+type CommentCreateData = {
+  authorId: string;
+  parentId?: string;
+  body: string;
+  createdTime: Date;
+};
+
+// ── Shared helpers ────────────────────────────────────────────────────────────
+
+function selectLikes(
+  db: DB,
+  commentIds: string[]
+): Map<string, CommentLiker[]> {
+  const byComment = new Map<string, CommentLiker[]>();
+  if (commentIds.length === 0) {
+    return byComment;
+  }
+  const rows = db
+    .select({
+      commentId: schema.commentLikes.commentId,
+      id: schema.guests.id,
+      name: schema.guests.name,
+      avatarUrl: schema.guests.avatarUrl,
+    })
+    .from(schema.commentLikes)
+    .innerJoin(schema.guests, eq(schema.commentLikes.guestId, schema.guests.id))
+    .where(inArray(schema.commentLikes.commentId, commentIds))
+    .orderBy(
+      asc(schema.commentLikes.createdTime),
+      asc(insertionOrder(schema.commentLikes))
+    )
+    .all();
+  for (const { commentId, id, name, avatarUrl } of rows) {
+    const likes = byComment.get(commentId) ?? [];
+    likes.push({ id, name, avatarUrl });
+    byComment.set(commentId, likes);
+  }
+  return byComment;
+}
+
+function listComments(db: DB, rows: CommentRow[]): Comment[] {
+  const likes = selectLikes(
+    db,
+    rows.map((r) => r.id)
+  );
+  return rows.map((row) => toComment(row, likes.get(row.id)));
+}
+
+function createdComment(db: DB, id: string, data: CommentCreateData): Comment {
+  const author = db
+    .select({ id: schema.guests.id, name: schema.guests.name })
+    .from(schema.guests)
+    .where(eq(schema.guests.id, data.authorId))
+    .get();
+  return {
+    id,
+    parentId: data.parentId ?? null,
+    body: data.body,
+    deleted: false,
+    createdTime: data.createdTime,
+    editedTime: null,
+    author: author ?? null,
+    likes: [],
+  };
+}
+
+// Ids are random nanoids, so they can't break a tie between two rows written in
+// the same millisecond; the implicit rowid is insertion order.
+function insertionOrder(table: SQLiteTable): SQL {
+  return sql`${table}.rowid`;
+}
+
+function toComment(row: CommentRow, likes: CommentLiker[] = []): Comment {
+  return {
+    id: row.id,
+    parentId: row.parentId,
+    body: row.body,
+    deleted: row.deleted,
+    createdTime: new Date(row.createdTime),
+    editedTime: row.editedTime ? new Date(row.editedTime) : null,
+    author:
+      row.authorId && row.authorName
+        ? { id: row.authorId, name: row.authorName }
+        : null,
+    likes,
+  };
+}
+
+// ── Scope-agnostic core ───────────────────────────────────────────────────────
+
 export class SqliteCommentsRepository implements CommentsRepository {
   constructor(private readonly db: DB) {}
 
-  private selectByProposal(proposalId: string): CommentRow[] {
-    return this.db
-      .select({
-        id: schema.comments.id,
-        parentId: schema.comments.parentId,
-        body: schema.comments.body,
-        deleted: schema.comments.deleted,
-        createdTime: schema.comments.createdTime,
-        editedTime: schema.comments.editedTime,
-        authorId: schema.guests.id,
-        authorName: schema.guests.name,
-      })
-      .from(schema.proposalComments)
-      .innerJoin(
-        schema.comments,
-        eq(schema.proposalComments.commentId, schema.comments.id)
-      )
-      .leftJoin(schema.guests, eq(schema.comments.authorId, schema.guests.id))
-      .where(eq(schema.proposalComments.proposalId, proposalId))
-      .orderBy(
-        asc(schema.comments.createdTime),
-        asc(insertionOrder(schema.comments))
-      )
-      .all();
-  }
-
-  private selectLikes(commentIds: string[]): Map<string, CommentLiker[]> {
-    const byComment = new Map<string, CommentLiker[]>();
-    if (commentIds.length === 0) {
-      return byComment;
-    }
-    const rows = this.db
-      .select({
-        commentId: schema.commentLikes.commentId,
-        id: schema.guests.id,
-        name: schema.guests.name,
-        avatarUrl: schema.guests.avatarUrl,
-      })
-      .from(schema.commentLikes)
-      .innerJoin(
-        schema.guests,
-        eq(schema.commentLikes.guestId, schema.guests.id)
-      )
-      .where(inArray(schema.commentLikes.commentId, commentIds))
-      .orderBy(
-        asc(schema.commentLikes.createdTime),
-        asc(insertionOrder(schema.commentLikes))
-      )
-      .all();
-    for (const { commentId, id, name, avatarUrl } of rows) {
-      const likes = byComment.get(commentId) ?? [];
-      likes.push({ id, name, avatarUrl });
-      byComment.set(commentId, likes);
-    }
-    return byComment;
-  }
-
-  async listByProposal(proposalId: string): Promise<Comment[]> {
-    const rows = this.selectByProposal(proposalId);
-    const likes = this.selectLikes(rows.map((r) => r.id));
-    return rows.map((row) => toComment(row, likes.get(row.id)));
-  }
-
-  async findById(id: string): Promise<Comment | undefined> {
+  async findById(commentId: string): Promise<Comment | undefined> {
     const row = this.db
-      .select({
-        id: schema.comments.id,
-        parentId: schema.comments.parentId,
-        body: schema.comments.body,
-        deleted: schema.comments.deleted,
-        createdTime: schema.comments.createdTime,
-        editedTime: schema.comments.editedTime,
-        authorId: schema.guests.id,
-        authorName: schema.guests.name,
-      })
+      .select(commentColumns)
       .from(schema.comments)
       .leftJoin(schema.guests, eq(schema.comments.authorId, schema.guests.id))
-      .where(eq(schema.comments.id, id))
+      .where(eq(schema.comments.id, commentId))
       .get();
-    return row && toComment(row, this.selectLikes([id]).get(id));
-  }
-
-  async findProposalId(commentId: string): Promise<string | undefined> {
-    return this.db
-      .select({ proposalId: schema.proposalComments.proposalId })
-      .from(schema.proposalComments)
-      .where(eq(schema.proposalComments.commentId, commentId))
-      .get()?.proposalId;
-  }
-
-  async createForProposal(data: {
-    proposalId: string;
-    authorId: string;
-    parentId?: string;
-    body: string;
-    createdTime: Date;
-  }): Promise<Comment> {
-    const id = nanoid();
-    this.db.transaction((tx) => {
-      tx.insert(schema.comments)
-        .values({
-          id,
-          authorId: data.authorId,
-          parentId: data.parentId ?? null,
-          body: data.body,
-          createdTime: data.createdTime.toISOString(),
-        })
-        .run();
-      tx.insert(schema.proposalComments)
-        .values({ commentId: id, proposalId: data.proposalId })
-        .run();
-    });
-    const author = this.db
-      .select({ id: schema.guests.id, name: schema.guests.name })
-      .from(schema.guests)
-      .where(eq(schema.guests.id, data.authorId))
-      .get();
-    return {
-      id,
-      parentId: data.parentId ?? null,
-      body: data.body,
-      deleted: false,
-      createdTime: data.createdTime,
-      editedTime: null,
-      author: author ?? null,
-      likes: [],
-    };
+    return (
+      row && toComment(row, selectLikes(this.db, [commentId]).get(commentId))
+    );
   }
 
   async toggleLike(data: {
@@ -250,24 +239,89 @@ export class SqliteCommentsRepository implements CommentsRepository {
   }
 }
 
-// Ids are random nanoids, so they can't break a tie between two rows written in
-// the same millisecond; the implicit rowid is insertion order.
-function insertionOrder(table: SQLiteTable): SQL {
-  return sql`${table}.rowid`;
+// ── Per-subject repositories ─────────────────────────────────────────────────
+
+type Tx = Parameters<Parameters<DB["transaction"]>[0]>[0];
+
+// A subject's join table, reduced to what attaching and looking up need.
+// `attach` is a callback rather than the table itself so each insert stays
+// typed against its own column names.
+type CommentJoin = {
+  table: SQLiteTable;
+  commentId: AnySQLiteColumn;
+  subjectId: AnySQLiteColumn;
+  attach: (tx: Tx, commentId: string, subjectId: string) => void;
+};
+
+const JOINS: Record<"proposal", CommentJoin> = {
+  proposal: {
+    table: schema.proposalComments,
+    commentId: schema.proposalComments.commentId,
+    subjectId: schema.proposalComments.proposalId,
+    attach: (tx, commentId, proposalId) =>
+      tx
+        .insert(schema.proposalComments)
+        .values({ commentId, proposalId })
+        .run(),
+  },
+};
+
+class SqliteSubjectCommentsRepository implements SubjectCommentsRepository {
+  constructor(
+    private readonly db: DB,
+    private readonly join: CommentJoin
+  ) {}
+
+  async list(subjectId: string): Promise<Comment[]> {
+    const rows = this.db
+      .select(commentColumns)
+      .from(this.join.table)
+      .innerJoin(schema.comments, eq(this.join.commentId, schema.comments.id))
+      .leftJoin(schema.guests, eq(schema.comments.authorId, schema.guests.id))
+      .where(eq(this.join.subjectId, subjectId))
+      .orderBy(
+        asc(schema.comments.createdTime),
+        asc(insertionOrder(schema.comments))
+      )
+      .all();
+    return listComments(this.db, rows);
+  }
+
+  async findSubjectId(commentId: string): Promise<string | undefined> {
+    return this.db
+      .select({ subjectId: this.join.subjectId })
+      .from(this.join.table)
+      .where(eq(this.join.commentId, commentId))
+      .get()?.subjectId as string | undefined;
+  }
+
+  async create(data: {
+    subjectId: string;
+    authorId: string;
+    parentId?: string;
+    body: string;
+    createdTime: Date;
+  }): Promise<Comment> {
+    const id = nanoid();
+    this.db.transaction((tx) => {
+      tx.insert(schema.comments)
+        .values({
+          id,
+          authorId: data.authorId,
+          parentId: data.parentId ?? null,
+          body: data.body,
+          createdTime: data.createdTime.toISOString(),
+        })
+        .run();
+      this.join.attach(tx, id, data.subjectId);
+    });
+    return createdComment(this.db, id, data);
+  }
 }
 
-function toComment(row: CommentRow, likes: CommentLiker[] = []): Comment {
-  return {
-    id: row.id,
-    parentId: row.parentId,
-    body: row.body,
-    deleted: row.deleted,
-    createdTime: new Date(row.createdTime),
-    editedTime: row.editedTime ? new Date(row.editedTime) : null,
-    author:
-      row.authorId && row.authorName
-        ? { id: row.authorId, name: row.authorName }
-        : null,
-    likes,
-  };
+export function sqliteSubjectCommentsRepository(
+  db: DB,
+  subject: keyof typeof JOINS
+): SubjectCommentsRepository {
+  return new SqliteSubjectCommentsRepository(db, JOINS[subject]);
 }
